removed unneeded complexity
Basically, I made it easier to make it so a block can have their own random tick logic while also being able to have random tick logic from their superclass. This made implementing other things a bit more annoying and also introduced more code. And also no vanilla block ever did something that requires this so there really was no reason to have this and i removed it. This kind of logic is however still in entities because vanilla entities do use it. An example: Grow up logic for passiveentities and restock logic for villagers which are part of passiveentities.