"Remember Block Positions" performance option
Before it took on average 2 ms to loop through a single chunk and takes 400 ms for 189 chunks.
With the new option turned on its 0 ms on average and takes like 17 ms for 189 chunks.
Yea. These results are obvious if you know the previous method would loop through 98304 blocks per chunk instead of the required blocks. But wow that's a lot of performance we've been missing out on!

This option is experimental because I am not yet confident with the code I just wrote.