Remove cardinal components dependency
Big changes! The main reason for this commit is because it was a fabric only mod and I wanted to port this mod to (neo)forge in the future. Another smaller reason was because it was causing a bit of warning spam for the users that uninstalled my mod but still had another mod that contained cardinal components. This has been fixed in the newer versions but I my mod is also supporting old versions of Minecraft. It was also offering a lot more than what I actually needed. I only needed to save/load data for the server to read and that's it.